Add another BugFix: Illegal Access in GC
As discussed on the mailing list (`PO: Illegal memory access in GC`),
the garbage collector will read from IO address -4 when it encounters a
NIL pointer and tries to read the metadata preceding the pointer. By
default, IO address -4 is unused and will return 0, but better to avoid
these accesses if possible.