Fix marking sectors used by LinkedExtensionTable
The code contained two bugs. First, as the linked extension table is a
linked list, it contains only 255 instead of 256 sector entries. Second,
the wrong loop variable was used to determine whether a sector is part
of the normal extension table or the linked extension table.

As a result, when a file larger than 3MB is present, some sectors
starting at 64KB of it were not marked as used and overwritten by the
next file.