Work around spurious freeze in GitHub workflow
Seems that System.Free can still cause memory corruption in the base
Oberon version that is used by the CommandLine emulator.

Work around it by restarting the emulator instead of relying on
System.Free.

Also limit the runtime for the GitHub workflow to 10 minutes (which is
plenty) to spot these kinds of errors faster.