Clean up some quirks
Not having a way to automatically run a command at startup yielded some
creative solutions, which can be removed now.

Also modify CommandLineDefragger to use a hook in the normal Defragger
instead of a patched module. Makes it easier to remove
CommandLineDefragger before finally defragmenting the disk images.

Last but not least, fix broken Clipboard and PCLink in DebugWithRescue
image's rescue mode.