Add MinimalFileystem modification
I wrote this modification about a year ago, and also believed I had
committed it, but it seems like I forgot it and just let it sit on my
disk. Anyway, here it is (tested that it still works). As it is not
included in the disk images anyway, no big deal.