UTF8CharsetLite: Fix font memory corruption
When a font contains large glyphs in the ASCII range and is loaded for
use with a non-Unicode viewer, the glyphs can overflow LargeFontDesc
RECORD size and corrupt memory. Add a new HugeFontDesc which is 32KB
larger than LargeFontDesc, and in case this is not sufficient, prefer
creating an ASSERT Trap to corrupting memory.